server/requirements.txt
fastapi==0.100.0
langchain==0.3.16
langchain-community==0.3.16
langchain-core==0.3.32
langchain-openai==0.3.2
langchain-text-splitters==0.3.5
langchainhub==0.1.21
moviepy==2.1.2
openai==1.60.2
pydub==0.25.1
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.31.0
uvicorn==0.23.1
